Refactoring code - table names and filenames should be the same
